+++ R/mark-package.R
@@ -12,6 +12,7 @@ op.mark <- list( # nolint: object_name_linter.
   # control for check_interactive() to return interactive() or TRUE
   mark.check_interactive = TRUE,
   mark.note.width        = NULL,
+  mark.todos.ext         = c("R", "Rmd", "qmd", "md", "py"),
   mark.todos.force       = FALSE,
 
   mark.days_in_month     = 30,